Promotion Function created
function takes a piece object and integer for promotion type and creates
a piece objects for the corresponding selection and returns the newly
created object. I also added a elseif statement in the OnMove() function
to check to see if the location is occupied and if the location is the
edge on the board. This statement will be changed in the future when I
have a better understanding of the program flow