New attribute in Piece class and Icons updated on promotion
I added a new attribute to the Piece class that allows for the easy
access of the drawable resources for the different icons. This allowed
for me to update the Icon of the Pawn when it is promoted to another
Piece. Currently the pieces are not assigned to a player when they are
created and can remove any piece on the board. I am also working on
getting the color to change to the appropriate team color.